+++ react/utils/stringUtils.ts
@@ -1,0 +1,3 @@
+export const removePrefixRegex = (str, prefix) => {
+  return str.replace(new RegExp(`^${prefix}`), '')
+}