+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Fix secure domain pdp url issue
+
 ## [2.0.3] - 2024-07-10
 
 ### Fixed